refactor(cli): remove legacy bin lib shims
