+++ .agents/skills/nemoclaw-maintainer-day/RISKY-AREAS.md
@@ -5,7 +5,7 @@ PRs touching these areas need tests before approval.
 | Area | Key paths |
 |------|-----------|
 | Installer / bootstrap shell | `install.sh`, `setup.sh`, `brev-setup.sh`, `scripts/*.sh` |
-| Onboarding / host glue | `bin/lib/onboard.js`, `bin/*.js` |
+| Onboarding / host glue | `src/lib/onboard.ts`, `bin/nemoclaw.js`, `scripts/*.sh` |
 | Sandbox / policy / SSRF | `nemoclaw/src/blueprint/`, `nemoclaw-blueprint/`, policy presets |
 | Workflow / enforcement | `.github/workflows/`, prek hooks, DCO, signing, version/tag flows |
 | Credentials / inference / network | credential helpers, inference provider routing, approval flows |
