--- bin/lib/debug.js
@@ -1,4 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-module.exports = require("../../dist/lib/debug");
