--- bin/lib/onboard.js
@@ -1,65 +1,0 @@
-// @ts-nocheck
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-const mod = require("../../dist/lib/onboard");
-module.exports = {
-  buildProviderArgs: mod.buildProviderArgs,
-  buildSandboxConfigSyncScript: mod.buildSandboxConfigSyncScript,
-  compactText: mod.compactText,
-  copyBuildContextDir: mod.copyBuildContextDir,
-  classifySandboxCreateFailure: mod.classifySandboxCreateFailure,
-  createSandbox: mod.createSandbox,
-  formatEnvAssignment: mod.formatEnvAssignment,
-  getFutureShellPathHint: mod.getFutureShellPathHint,
-  getGatewayStartEnv: mod.getGatewayStartEnv,
-  getGatewayReuseState: mod.getGatewayReuseState,
-  getNavigationChoice: mod.getNavigationChoice,
-  getSandboxInferenceConfig: mod.getSandboxInferenceConfig,
-  getInstalledOpenshellVersion: mod.getInstalledOpenshellVersion,
-  getBlueprintMinOpenshellVersion: mod.getBlueprintMinOpenshellVersion,
-  versionGte: mod.versionGte,
-  getRequestedModelHint: mod.getRequestedModelHint,
-  getRequestedProviderHint: mod.getRequestedProviderHint,
-  getStableGatewayImageRef: mod.getStableGatewayImageRef,
-  getResumeConfigConflicts: mod.getResumeConfigConflicts,
-  isGatewayHealthy: mod.isGatewayHealthy,
-  hasStaleGateway: mod.hasStaleGateway,
-  getRequestedSandboxNameHint: mod.getRequestedSandboxNameHint,
-  getResumeSandboxConflict: mod.getResumeSandboxConflict,
-  getSandboxReuseState: mod.getSandboxReuseState,
-  getSandboxStateFromOutputs: mod.getSandboxStateFromOutputs,
-  getPortConflictServiceHints: mod.getPortConflictServiceHints,
-  classifyValidationFailure: mod.classifyValidationFailure,
-  isSandboxReady: mod.isSandboxReady,
-  isLoopbackHostname: mod.isLoopbackHostname,
-  normalizeProviderBaseUrl: mod.normalizeProviderBaseUrl,
-  onboard: mod.onboard,
-  onboardSession: mod.onboardSession,
-  printSandboxCreateRecoveryHints: mod.printSandboxCreateRecoveryHints,
-  providerExistsInGateway: mod.providerExistsInGateway,
-  parsePolicyPresetEnv: mod.parsePolicyPresetEnv,
-  pruneStaleSandboxEntry: mod.pruneStaleSandboxEntry,
-  repairRecordedSandbox: mod.repairRecordedSandbox,
-  recoverGatewayRuntime: mod.recoverGatewayRuntime,
-  resolveDashboardForwardTarget: mod.resolveDashboardForwardTarget,
-  startGatewayForRecovery: mod.startGatewayForRecovery,
-  runCaptureOpenshell: mod.runCaptureOpenshell,
-  setupInference: mod.setupInference,
-  setupMessagingChannels: mod.setupMessagingChannels,
-  setupNim: mod.setupNim,
-  isInferenceRouteReady: mod.isInferenceRouteReady,
-  isOpenclawReady: mod.isOpenclawReady,
-  arePolicyPresetsApplied: mod.arePolicyPresetsApplied,
-  presetsCheckboxSelector: mod.presetsCheckboxSelector,
-  setupPoliciesWithSelection: mod.setupPoliciesWithSelection,
-  summarizeCurlFailure: mod.summarizeCurlFailure,
-  summarizeProbeFailure: mod.summarizeProbeFailure,
-  hasResponsesToolCall: mod.hasResponsesToolCall,
-  upsertProvider: mod.upsertProvider,
-  hydrateCredentialEnv: mod.hydrateCredentialEnv,
-  pruneKnownHostsEntries: mod.pruneKnownHostsEntries,
-  shouldIncludeBuildContextPath: mod.shouldIncludeBuildContextPath,
-  writeSandboxConfigSyncFile: mod.writeSandboxConfigSyncFile,
-  patchStagedDockerfile: mod.patchStagedDockerfile,
-};
