--- bin/lib/preflight.js
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-//
-// Thin re-export shim — the implementation lives in src/lib/preflight.ts,
-// compiled to dist/lib/preflight.js.
-
-module.exports = require("../../dist/lib/preflight");
