--- bin/lib/services.js
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Thin CJS shim — implementation lives in src/lib/services.ts
-module.exports = require("../../dist/lib/services");
