+++ package.json
@@ -32,6 +32,7 @@
   ],
   "files": [
     "bin/",
+    "dist/",
     "nemoclaw/dist/",
     "nemoclaw/openclaw.plugin.json",
     "nemoclaw/package.json",
