+++ src/lib/local-inference.ts
@@ -7,7 +7,7 @@
  */
 
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-const { shellQuote } = require("../../bin/lib/runner");
+const { shellQuote } = require("./runner");
 
 export const HOST_GATEWAY_URL = "http://host.openshell.internal";
 export const CONTAINER_REACHABILITY_IMAGE = "curlimages/curl:8.10.1";
