+++ src/lib/model-prompts.ts
@@ -5,9 +5,9 @@ import { CLOUD_MODEL_OPTIONS } from "./inference-config";
 import { isSafeModelId } from "./validation";
 import { validateNvidiaEndpointModel } from "./provider-models";
 
-// credentials.js is CJS.
+// credentials.ts still uses CommonJS-style exports.
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-const { getCredential, prompt } = require("../../bin/lib/credentials");
+const { getCredential, prompt } = require("./credentials");
 
 export const BACK_TO_SELECTION = "__NEMOCLAW_BACK_TO_SELECTION__";
