+++ src/lib/preflight.ts
@@ -14,9 +14,9 @@ import net from "node:net";
 import os from "node:os";
 import path from "node:path";
 
-// runner.js is CJS — use require so we don't pull it into the TS build.
+// runner.ts still uses CommonJS-style exports — use require here.
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-const { runCapture } = require("../../bin/lib/runner");
+const { runCapture } = require("./runner");
 
 // ── Types ────────────────────────────────────────────────────────
