+++ src/lib/provider-models.test.ts
@@ -11,7 +11,7 @@ import {
   validateAnthropicModel,
   validateNvidiaEndpointModel,
   validateOpenAiLikeModel,
-} from "./provider-models";
+} from "../../dist/lib/provider-models";
 
 describe("provider model helpers", () => {
   it("fetches NVIDIA endpoint model ids", () => {
