+++ test/e2e/test-onboard-repair.sh
@@ -115,7 +115,7 @@ fi
 node -e '
 const { saveCredential } = require(process.argv[1]);
 saveCredential("NVIDIA_API_KEY", process.argv[2]);
-' "$REPO/bin/lib/credentials.js" "$RESTORE_API_KEY"
+' "$REPO/dist/lib/credentials.js" "$RESTORE_API_KEY"
 pass "Stored NVIDIA_API_KEY in ~/.nemoclaw/credentials.json for resume hydration"
 
 # ══════════════════════════════════════════════════════════════════
