+++ test/e2e/test-onboard-resume.sh
@@ -136,7 +136,7 @@ fi
 node -e '
 const { saveCredential } = require(process.argv[1]);
 saveCredential("NVIDIA_API_KEY", process.argv[2]);
-' "$REPO/bin/lib/credentials.js" "$RESTORE_API_KEY"
+' "$REPO/dist/lib/credentials.js" "$RESTORE_API_KEY"
 pass "Stored NVIDIA_API_KEY in ~/.nemoclaw/credentials.json for resume hydration"
 
 # ══════════════════════════════════════════════════════════════════
