+++ test/e2e/test-telegram-injection.sh
@@ -345,7 +345,7 @@ info "T6: Testing SANDBOX_NAME with shell metacharacters..."
 # alphanumeric with optional internal hyphens, max 63 chars.
 # Test by running the validation directly via node.
 t6_result=$(cd "$REPO" && node -e "
-  const { validateName } = require('./bin/lib/runner');
+  const { validateName } = require('./dist/lib/runner');
   try {
     validateName('foo;rm -rf /', 'SANDBOX_NAME');
     console.log('ACCEPTED');
@@ -364,7 +364,7 @@ fi
 info "T7: Testing SANDBOX_NAME with leading hyphen (option injection)..."
 
 t7_result=$(cd "$REPO" && node -e "
-  const { validateName } = require('./bin/lib/runner');
+  const { validateName } = require('./dist/lib/runner');
   try {
     validateName('--help', 'SANDBOX_NAME');
     console.log('ACCEPTED');
@@ -383,7 +383,7 @@ fi
 # backticks and $() in double-quoted node -e strings.
 for invalid_name in '$(whoami)' '`id`' 'foo bar' '../etc/passwd' 'UPPERCASE'; do
   t_result=$(cd "$REPO" && node -e "
-    const { validateName } = require('./bin/lib/runner');
+    const { validateName } = require('./dist/lib/runner');
     try {
       validateName(process.argv[1], 'SANDBOX_NAME');
       console.log('ACCEPTED');
