+++ test/sandbox-build-context.test.ts
@@ -11,7 +11,7 @@ import {
   collectBuildContextStats,
   stageLegacySandboxBuildContext,
   stageOptimizedSandboxBuildContext,
-} from "../bin/lib/sandbox-build-context";
+} from "../dist/lib/sandbox-build-context";
 
 describe("sandbox build context staging", () => {
   it("optimized staging excludes blueprint .venv and extra scripts while preserving required files", () => {
